# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Image Hosting Service API
  version: 1.0.0
  description: |
    REST API for uploading, managing, searching, and securely sharing images.

    Uploaded images are private by default. Original files and thumbnails are
    stored in a private Cloudflare R2 bucket and delivered through protected API
    endpoints. Gemini analyzes images asynchronously and generates searchable
    objects, tags, and colors.

    Authentication uses an opaque database-backed SESSION cookie. Authenticated
    state-changing requests also require the X-CSRF-TOKEN header.
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /images:
    post:
      tags:
        - Images
      summary: Upload an image
      description: |
        Uploads a JPEG, PNG, or WebP image up to 10 MB.

        Every uploaded image is private by default. Thumbnail generation is
        completed during upload, and Gemini analysis continues asynchronously.

        The response contains image metadata and relative thumbnail and content
        URLs. It does not contain image bytes.
      operationId: uploadImage
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: JPEG, PNG, or WebP image up to 10 MB
      responses:
        "201":
          description: Image uploaded as private
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageResponse"
        "400":
          $ref: "#/components/responses/InvalidImage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedImageType"

    get:
      tags:
        - Images
      summary: Browse or search public images
      description: |
        Returns public images from newest to oldest.

        Each item uses ImageResponse and contains image metadata plus relative
        thumbnail and content URLs. The response does not contain image bytes.

        Gallery and search clients should display thumbnailUrl. They should
        request contentUrl only after the user explicitly opens the full image.

        When `q` is supplied, the query searches AI-generated objects, tags,
        and colors using case-insensitive partial matching. All supplied
        filters are combined with AND. Private images are never included.
      operationId: getPublicImages
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/ContentTypeFilter"
        - $ref: "#/components/parameters/ColorFilter"
        - $ref: "#/components/parameters/TaggingStatusFilter"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Page of public image metadata with thumbnail and content URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagePageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /images/mine:
    get:
      tags:
        - Images
      summary: List owned images
      description: |
        Returns public and private images owned by the authenticated user from
        newest to oldest.

        Each item uses ImageResponse and contains image metadata plus relative
        thumbnail and content URLs. The response does not contain image bytes.

        Clients should display thumbnailUrl in the image library and request
        contentUrl only after the user explicitly opens the full image.

        Search is limited to the authenticated user's images. All supplied
        filters are combined with AND.
      operationId: getMyImages
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/ContentTypeFilter"
        - $ref: "#/components/parameters/ColorFilter"
        - $ref: "#/components/parameters/TaggingStatusFilter"
        - $ref: "#/components/parameters/VisibilityFilter"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Page of owned image metadata with thumbnail and content URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagePageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /images/{imageId}:
    parameters:
      - $ref: "#/components/parameters/ImageId"

    get:
      tags:
        - Images
      summary: Get image metadata
      description: |
        Returns metadata and relative content and thumbnail URLs for an
        accessible image. This response does not contain image bytes.

        Public images are available to everyone. Private images are available
        only to their owner. An inaccessible private image is reported as not
        found.
      operationId: getImage
      security:
        - {}
        - sessionCookie: []
      responses:
        "200":
          description: Accessible image metadata with thumbnail and content URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageResponse"
        "404":
          $ref: "#/components/responses/ImageNotFound"

    patch:
      tags:
        - Images
      summary: Change image visibility
      description: |
        Makes an owned image public or private.

        Only the image owner can change its visibility. The response contains
        updated image metadata and relative image URLs.
      operationId: updateImageVisibility
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateImageVisibilityRequest"
      responses:
        "200":
          description: Visibility updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/ImageNotFound"

    delete:
      tags:
        - Images
      summary: Delete an image
      description: |
        Permanently deletes an owned image, including its database record,
        original object, thumbnail, and sharing links.
      operationId: deleteImage
      security:
        - sessionCookie: []
          csrfToken: []
      responses:
        "204":
          description: Image deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/ImageNotFound"
  /images/{imageId}/content:
    get:
      tags:
        - Images
      summary: Stream the original image
      description: |
        Returns the original full-size image bytes.

        Clients should request this endpoint only after the user explicitly
        chooses to view the full image. Galleries, search results, and image
        libraries should use the thumbnail endpoint instead.

        Public images are available to everyone. Private images are available
        only to their owner.
      operationId: getImageContent
      security:
        - {}
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ImageId"
      responses:
        "200":
          description: Original full-size image content
          headers:
            Content-Disposition:
              $ref: "#/components/headers/ContentDisposition"
            X-Content-Type-Options:
              $ref: "#/components/headers/NoSniff"
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/ImageNotFound"
  /images/{imageId}/thumbnail:
    get:
      tags:
        - Images
      summary: Stream the image thumbnail
      description: |
        Returns the generated 100 × 100 PNG thumbnail. Thumbnails are always
        PNG regardless of whether the original image is JPEG, PNG, or WebP.

        This endpoint should be used when displaying public galleries, search
        results, and the authenticated user's image library.

        Public-image thumbnails are available to everyone. Private-image
        thumbnails are available only to their owner.
      operationId: getImageThumbnail
      security:
        - {}
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ImageId"
      responses:
        "200":
          description: Generated 100 × 100 thumbnail content
          headers:
            Content-Disposition:
              $ref: "#/components/headers/ContentDisposition"
            X-Content-Type-Options:
              $ref: "#/components/headers/NoSniff"
          content:
            image/png:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/ImageNotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: Opaque session identifier stored by Spring Session JDBC

    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: CSRF token returned by GET /auth/csrf
  parameters:
    ImageId:
      name: imageId
      in: path
      required: true
      description: |
        Stable public UUID of the image. The internal numeric database ID is
        never exposed through the API.
      schema:
        type: string
        format: uuid
      example: c98237ef-192c-4fd5-b2e4-7e0bd5a68253
...
    SearchQuery:
      name: q
      in: query
      required: false
      description: |
        Optional case-insensitive partial search across AI-generated objects,
        tags, and colors.
      schema:
        type: string
        minLength: 1
        maxLength: 100
      example: sunset
    ContentTypeFilter:
      name: contentType
      in: query
      required: false
      description: Return only images with this original content type
      schema:
        type: string
        enum:
          - image/jpeg
          - image/png
          - image/webp
      example: image/webp
    ColorFilter:
      name: color
      in: query
      required: false
      description: Case-insensitive exact match against AI-detected colors
      schema:
        type: string
        minLength: 1
        maxLength: 100
      example: orange
    TaggingStatusFilter:
      name: taggingStatus
      in: query
      required: false
      description: Return only images with this AI-tagging status
      schema:
        $ref: "#/components/schemas/TaggingStatus"
      example: COMPLETED
    VisibilityFilter:
      name: isPublic
      in: query
      required: false
      description: Return only public or private images owned by the authenticated user
      schema:
        type: boolean
      example: false
    Page:
      name: page
      in: query
      required: false
      description: Zero-based page number
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: Number of images per page
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
  headers:
...
    ContentDisposition:
      description: Inline content disposition with the sanitized filename
      schema:
        type: string
        example: inline; filename="image.webp"
    NoSniff:
      description: Prevents browser content-type sniffing
      schema:
        type: string
        enum:
          - nosniff
  schemas:
...
    ImageTags:
      type: object
      additionalProperties: false
      required:
        - objects
        - tags
        - colors
      properties:
        objects:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            minLength: 1
          example:
            - ocean
            - cloud
        tags:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            minLength: 1
          example:
            - sunset
            - nature
            - evening
        colors:
          type: array
          maxItems: 3
          uniqueItems: true
          items:
            $ref: "#/components/schemas/ImageColor"
          example:
            - ORANGE
            - BLUE
            - PURPLE
    ImageColor:
      type: string
      enum:
        - BLACK
        - WHITE
        - RED
        - GREEN
        - YELLOW
        - BLUE
        - BROWN
        - ORANGE
        - PINK
        - PURPLE
        - GRAY
    TaggingStatus:
      type: string
      enum:
        - PENDING
        - PROCESSING
        - COMPLETED
        - FAILED
    ImageResponse:
      type: object
      description: |
        Image metadata shared by upload, gallery, search, owner-library,
        visibility-update, and individual-image responses.

        The response contains relative API URLs. It does not contain thumbnail
        or original image bytes.

        Clients should use thumbnailUrl in galleries, search results, and image
        libraries. Clients should request contentUrl only after the user
        explicitly opens the full image.
      additionalProperties: false
      required:
        - id
        - originalFilename
        - contentType
        - sizeBytes
        - width
        - height
        - isPublic
        - taggingStatus
        - contentUrl
        - thumbnailUrl
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: |
            Stable public image UUID. This is different from the internal
            numeric database ID.
          example: c98237ef-192c-4fd5-b2e4-7e0bd5a68253

        originalFilename:
          type: string
          maxLength: 255
          description: Sanitized original filename
          example: sunset.webp

        contentType:
          type: string
          enum:
            - image/jpeg
            - image/png
            - image/webp
          example: image/webp

        sizeBytes:
          type: integer
          format: int64
          minimum: 1
          maximum: 10485760
          description: Original image size in bytes
          example: 284391

        width:
          type: integer
          minimum: 1
          description: Original image width in pixels
          example: 1920

        height:
          type: integer
          minimum: 1
          description: Original image height in pixels
          example: 1080

        isPublic:
          type: boolean
          description: Whether the image can be accessed without authentication
          example: false

        aiTags:
          description: AI-generated image metadata, or null while unavailable
          anyOf:
            - $ref: "#/components/schemas/ImageTags"
            - type: "null"

        taggingStatus:
          $ref: "#/components/schemas/TaggingStatus"

        taggingStartedAt:
          type: [string, "null"]
          format: date-time
          example: "2026-07-24T12:00:01Z"

        taggingCompletedAt:
          type: [string, "null"]
          format: date-time
          example: "2026-07-24T12:00:04Z"

        taggingDurationMs:
          type: [integer, "null"]
          format: int64
          minimum: 0
          example: 3000

        taggingError:
          type: [string, "null"]
          description: Sanitized processing failure information
          example: null

        contentUrl:
          type: string
          description: |
            Relative API URL for streaming the original full-size image.
            Clients should request it only after the user explicitly opens
            the image.
          example: /api/v1/images/c98237ef-192c-4fd5-b2e4-7e0bd5a68253/content

        thumbnailUrl:
          type: string
          description: |
            Relative API URL for the generated thumbnail. Clients should use
            it in galleries, search results, and image-library lists.
          example: /api/v1/images/c98237ef-192c-4fd5-b2e4-7e0bd5a68253/thumbnail

        createdAt:
          type: string
          format: date-time
          example: "2026-07-24T12:00:00Z"

        updatedAt:
          type: [string, "null"]
          format: date-time
          example: "2026-07-24T12:00:04Z"
    ImagePageResponse:
      type: object
      description: |
        Paginated image metadata used by the public gallery, search results,
        and authenticated user's image library.

        Each item uses ImageResponse and contains URLs only. Image bytes are
        retrieved separately through the thumbnail or content endpoints.
      additionalProperties: false
      required:
        - images
        - page
        - size
        - totalElements
        - totalPages
      properties:
        images:
          type: array
          description: |
            Image metadata ordered from newest to oldest. Clients should use
            each item's thumbnailUrl when rendering the list.
          items:
            $ref: "#/components/schemas/ImageResponse"

        page:
          type: integer
          minimum: 0
          example: 0

        size:
          type: integer
          minimum: 1
          maximum: 50
          example: 20

        totalElements:
          type: integer
          format: int64
          minimum: 0
          example: 42

        totalPages:
          type: integer
          minimum: 0
          example: 3
    UpdateImageVisibilityRequest:
      type: object
      additionalProperties: false
      required:
        - isPublic
      properties:
        isPublic:
          type: boolean
          description: New public visibility state for the owned image
          example: true
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-07-24T12:00:00Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Request validation failed
        path:
          type: string
          example: /api/v1/images
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            timestamp: "2026-07-24T12:00:00Z"
            status: 400
            error: Bad Request
            message: Request validation failed
            path: /api/v1/images
    InvalidImage:
      description: Image is missing, empty, corrupt, or otherwise invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication is required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: The request is not allowed or the CSRF token is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ImageNotFound:
      description: Image does not exist or is not accessible
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    PayloadTooLarge:
      description: Image exceeds the 10 MB limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnsupportedImageType:
      description: Only JPEG, PNG, and WebP images are supported
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"